Attempt to fix CI

+++ r/src/filesystem.cpp
@@ -474,7 +474,7 @@ cpp11::list fs___GcsFileSystem__options(const std::shared_ptr<fs::GcsFileSystem>
 
     metadata.names() = opts.default_metadata->keys();
 
-    for (size_t i = 0; i < opts.default_metadata->size(); i++) {
+    for (int64_t i = 0; i < opts.default_metadata->size(); i++) {
       metadata[i] = opts.default_metadata->value(i);
     }
